Better error messages
One problem with relying on reflection is that when something crashes you don't necessarily have a good idea of what caused it. Prior to this change I was just getting generic "null object reference." This now gives much more detailed information and helps you track down the root cause of the problem.